pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=4.3.0",
    "dependency-injector>=4.48.1",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-json-logger>=3.3.0",
    "sqlalchemy>=2.0.43",
    "user-agents>=2.2.0",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.10",
]

[tool.ruff]
target-version = "py313"
line-length = 72

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"*/views.py" = ["B008"]  # Allow FastAPI dependency injection patterns

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing"
pythonpath = ["."]
asyncio_mode = "auto"
